plate/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow
selenium==4.10.0
